// src/features/wishlists/components/create-form/interfaces/template-props.interface.ts
import React from 'react';

export interface TemplateProps {
  title: string;
  setTitle: (val: string) => void;
  expiresAt: string;
  setExpiresAt: (val: string) => void;
  allowGroupFunds: boolean;
  setAllowGroupFunds: (val: boolean) => void;
  isLoading: boolean;
  errorMsg: string | null;
  handleSubmit: (e: React.SyntheticEvent) => void;
  category: string;
  setCategory: (val: string) => void;
  customCategory: string;
  setCustomCategory: (val: string) => void;
  aiEnabled: boolean;
  setAiEnabled: (val: boolean) => void;
  webSearchEnabled: boolean;
  setWebSearchEnabled: (val: boolean) => void;
  autoRollover: boolean;
  setAutoRollover: (val: boolean) => void;
  globalAiEnabled: boolean;
  globalWebSearchEnabled: boolean;
  advancedOpen: boolean;
  setAdvancedOpen: React.Dispatch<React.SetStateAction<boolean>>;
  onCancel?: () => void;
}